verifyファイルをカテゴリごとにフォルダ分け
