+++ verify/yosupo/data_structure/line_add_get_min.cpp
@@ -1,6 +1,6 @@
 //@yosupo line_add_get_min
-#include "../../template.cpp"
-#include "../../Structure/LiChaoTree.cpp"
+#include "../../../template.cpp"
+#include "../../../Structure/LiChaoTree.cpp"
 struct query {
 	ll t;
 	ll a, b;
